Initial Commit
---Initial Commit---
Plugins included:
-JSimple-Slider, a very simple JQuery plugin for creating very simple
sliders.
-JSmooth-Scroll, a simple plugin to animate the transition between
elements in a document based on internal page links.
-JSocial-Buttons, Useful for social-media logos, when hovered over
transitions to the logo colours of the social media site in question.
-JBorder-to-background, simple plugin that animates the change of  the
colour of the background of a button/anything to its border colour while
also changing the text colour to the original background colour.

---Upcoming Plugins---
-Slider/Carousel with controls and pause on hover.

Also included are .min's of each plugin.